feat: Add charon groups as logical groupings of connections instead of nodes
This allows us to have multiple charon connections on the same node
provided they are not in the host network namespace. With it we're
adding a way to add annotations to charon pod so you can use ipman in
conjunction with other annotation based controllers such as our own
vlanman and multus.

fix: duplicating of configs across groups when 2 or more groups were on
the same node

feat: performance improvement to gOcaml

feat: allow the user to set the name of the interface

fix: xfrminjector docker container tag

fix: remove old TODO.md